Moved offline payments to separate service
