+++ Api/Controllers/AdministratorControllers/PaymentsController.cs
@@ -24,13 +24,15 @@ namespace HappyTravel.Edo.Api.Controllers.AdministratorControllers
     {
         public PaymentsController(IBookingPaymentService bookingPaymentService,
             IAdministratorContext administratorContext, ICounterpartyAccountService counterpartyAccountService, IAgencyAccountService agencyAccountService,
-            ICreditCardPaymentConfirmationService creditCardPaymentConfirmationService)
+            ICreditCardPaymentConfirmationService creditCardPaymentConfirmationService,
+            IBookingOfflinePaymentService offlinePaymentService)
         {
             _bookingPaymentService = bookingPaymentService;
             _administratorContext = administratorContext;
             _counterpartyAccountService = counterpartyAccountService;
             _agencyAccountService = agencyAccountService;
             _creditCardPaymentConfirmationService = creditCardPaymentConfirmationService;
+            _offlinePaymentService = offlinePaymentService;
         }
 
 
@@ -45,7 +47,7 @@ namespace HappyTravel.Edo.Api.Controllers.AdministratorControllers
         public async Task<IActionResult> CompleteOffline(int bookingId)
         {
             var (_, _, administrator, _) = await _administratorContext.GetCurrent();
-            var (_, isFailure, error) = await _bookingPaymentService.CompleteOffline(bookingId, administrator);
+            var (_, isFailure, error) = await _offlinePaymentService.CompleteOffline(bookingId, administrator);
             if (isFailure)
                 return BadRequest(ProblemDetailsBuilder.Build(error));
 
@@ -235,5 +237,6 @@ namespace HappyTravel.Edo.Api.Controllers.AdministratorControllers
         private readonly IBookingPaymentService _bookingPaymentService;
         private readonly ICounterpartyAccountService _counterpartyAccountService;
         private readonly ICreditCardPaymentConfirmationService _creditCardPaymentConfirmationService;
+        private readonly IBookingOfflinePaymentService _offlinePaymentService;
     }
 }
