+++ Api/Infrastructure/ServiceCollectionExtensions.cs
@@ -451,6 +451,7 @@ namespace HappyTravel.Edo.Api.Infrastructure
             services.AddTransient<IAgencyAccountService, AgencyAccountService>();
             services.AddTransient<IPaymentSettingsService, PaymentSettingsService>();
             services.AddTransient<IBookingPaymentService, BookingPaymentService>();
+            services.AddTransient<IBookingOfflinePaymentService, BookingOfflinePaymentService>();
             services.AddTransient<IBookingCreditCardPaymentService, BookingCreditCardPaymentService>();
             services.AddTransient<IBookingAccountPaymentService, BookingAccountPaymentService>();
             services.AddTransient<IBookingPaymentInfoService, BookingPaymentInfoService>();
